+++ src/typescript/fiber-sys/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @meshx-org/fiber-sys
 
+## 0.3.0
+
+### Minor Changes
+
+- test
+
+### Patch Changes
+
+- Updated dependencies
+  - @meshx-org/fiber-types@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes