+++ src/typescript/fiber-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @meshx-org/fiber-types
 
+## 0.3.0
+
+### Minor Changes
+
+- test
+
 ## 0.2.0
 
 ### Minor Changes